+++ janusgraph-core/src/main/java/org/janusgraph/util/stats/MetricManager.java
@@ -15,23 +15,18 @@
 package org.janusgraph.util.stats;
 
 import com.codahale.metrics.*;
-import com.codahale.metrics.ganglia.GangliaReporter;
 import com.codahale.metrics.graphite.Graphite;
 import com.codahale.metrics.graphite.GraphiteReporter;
 import com.google.common.base.Preconditions;
-import info.ganglia.gmetric4j.gmetric.GMetric;
-import info.ganglia.gmetric4j.gmetric.GMetric.UDPAddressingMode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.management.MBeanServer;
 import javax.management.MBeanServerFactory;
 import java.io.File;
-import java.io.IOException;
 import java.net.InetSocketAddress;
 import java.time.Duration;
 import java.util.List;
-import java.util.UUID;
 import java.util.concurrent.TimeUnit;
 
 
@@ -51,7 +46,6 @@ public enum MetricManager {
     private CsvReporter csvReporter           = null;
     private JmxReporter jmxReporter           = null;
     private Slf4jReporter slf4jReporter       = null;
-    private GangliaReporter gangliaReporter   = null;
     private GraphiteReporter graphiteReporter = null;
 
     /**
@@ -244,79 +238,6 @@ public enum MetricManager {
     }
 
     /**
-     * Create a {@link GangliaReporter} attached to the JanusGraph Metrics registry.
-     * <p>
-     * {@code groupOrHost} and {@code addressingMode} must be non-null. The
-     * remaining non-primitive arguments may be null. If {@code protocol31} is
-     * null, then true is assumed. Null values of {@code hostUUID} or
-     * {@code spoof} are passed into the {@link GMetric} constructor, which
-     * causes Ganglia to use its internal logic for generating a default UUID
-     * and default reporting hostname (respectively).
-     *
-     * @param groupOrHost
-     *            the multicast group or unicast hostname to which Ganglia
-     *            events are sent
-     * @param port
-     *            the port to which events are sent
-     * @param addressingMode
-     *            whether to send events with multicast or unicast
-     * @param ttl
-     *            multicast ttl (ignored for unicast)
-     * @param protocol31
-     *            true to use Ganglia protocol version 3.1, false to use 3.0
-     * @param hostUUID
-     *            uuid for the host
-     * @param spoof
-     *            override this machine's IP/hostname as it appears on the
-     *            Ganglia server
-     * @param reportInterval
-     *            time to wait before sending data to the ganglia
-     *            unicast host or multicast group
-     * @throws IOException
-     *             when a {@link GMetric} can't be instantiated using the
-     *             provided arguments
-     */
-    public synchronized void addGangliaReporter(String groupOrHost, int port,
-            UDPAddressingMode addressingMode, int ttl, Boolean protocol31,
-            UUID hostUUID, String spoof, Duration reportInterval) throws IOException {
-
-        Preconditions.checkNotNull(groupOrHost);
-        Preconditions.checkNotNull(addressingMode);
-
-        if (null != gangliaReporter) {
-            log.debug("Metrics GangliaReporter already active; not creating another");
-            return;
-        }
-
-        if (null == protocol31)
-            protocol31 = true;
-
-        GMetric ganglia = new GMetric(groupOrHost, port, addressingMode, ttl,
-                protocol31, hostUUID, spoof);
-
-        GangliaReporter.Builder b = GangliaReporter.forRegistry(getRegistry());
-
-        gangliaReporter = b.build(ganglia);
-        gangliaReporter.start(reportInterval.toMillis(), TimeUnit.MILLISECONDS);
-
-        log.info("Configured Ganglia Metrics reporter host={} interval={} port={} addrmode={} ttl={} proto31={} uuid={} spoof={}",
-            groupOrHost, reportInterval, port, addressingMode, ttl, protocol31, hostUUID, spoof);
-    }
-
-    /**
-     * Stop a {@link GangliaReporter} previously created by a call to
-     * {@link #addGangliaReporter(String, int, UDPAddressingMode, int, Boolean, UUID, String, Duration)}
-     * and release it for GC. Idempotent between calls to the associated add
-     * method. Does nothing before the first call to the associated add method.
-     */
-    public synchronized void removeGangliaReporter() {
-        if (null != gangliaReporter)
-            gangliaReporter.stop();
-
-        gangliaReporter = null;
-    }
-
-    /**
      * Create a {@link GraphiteReporter} attached to the JanusGraph Metrics registry.
      * <p>
      * If {@code prefix} is null, then Metrics's internal default prefix is used
@@ -375,7 +296,6 @@ public enum MetricManager {
         removeCsvReporter();
         removeJmxReporter();
         removeSlf4jReporter();
-        removeGangliaReporter();
         removeGraphiteReporter();
     }